refactor(backend): move jwt creation to dedicated file
